Update AutoDISM.bat
Sometimes we make little accidents. Sometimes we modify the wrong version of AutoDISM and add the condition checks to an older, more inferior version of AutoDISM, and want to roll the changes back. Sometimes, the github website can't perform what feels like the most basic of version controls, so instead of rolling the changes back to the previous commit, I have to make a new commit to cover up the old commit because I can't just roll the changes back. Sometimes I just don't understand most things, and the chunkiness of this website is one of them. Anways, I am *ROLLING BACK THE CHANGES TO THE PREVIOUS VERSION AND CHANGING THE COMMIT ID TO A NEW ONE BECAUSE I WANT TO*, and *I AM LEAVING THE FAILURE VERSION UP TO REFERENCE BACK TO BECAUSE I WANT A CONSTANT REMINDER OF MY FAILURES WITH WONDERFUL GREEN AND RED HIGHLIGHTING*. I am also *REMOVING THE ERROR MESSAGE HANDLING BECAUSE THAT WOULD LOOK NICER AND WE DEMAND NOTHING MORE THAN ABSOLUTE CHAOS BECAUSE SOMEONE FORGOT TO ADD THEM TO THE UPDATED PREVIOUS VERSION OF THE SCRIPT* and just, seriously Github, allow users to roll changes back from the website. It doesn't seem like it would be that hard, just add it in the three dots dropdown when viewing from a previous commit, or something.